requirements.txt
numpy
mpi4py
requests
responses
pytest
pytest-cov
scipy
tqdm
h5py
pandas
recordclass
cufflinks
googlemaps
matplotlib
plotly
PyYAML
scikit_learn
openpyxl
cartopy
geos
geopandas
geoplot
jupyter
utm
